server/requirements.txt
sanic
python-dotenv
anthropic
sanic-cors
tortoise-orm
aiomysql